propagate_memlets() extension
In propagate_memlets, if union_inner_edges = True, only add edges that have the same connector to the set that gets propagated as well. If the edge to propagate does not have a connector yet, jump back to default behaviour. 
This fixes volume issues, to see this in action, again launch jacobi2d in the polybench samples