CUDA events:
If the source s of an edge e has a CUDA stream but the destination d does not, we need to check if the destination md of the e's memlet path is different than  d. If it is and md has a CUDA stream we need to compare the CUDA streams between s and md. If they are not identical there needs to be a CUDA event on the edge.