do not hardcode c++ compiler in test script
The test_all.sh script currently runs cpp tests using g++.
This is not good for the following reasons:
* During normal DaCe compilation we use cmake (and thus whatever compiler cmake picks up).
* We don't check if g++ is available.

This change uses whatever the user set as CXX env var. Cmake also uses CXX when it is set.
Thus if a user sets CXX, he will use the same compiler for tests and during dace compilation.
If CXX is not set we fall back to g++ as the hard-coded compiler.

The test script also prints the current progress before each test now.