[bug] Missing symbols for data access on inter-state edge
After uplift to dace v0.15, one SDFG which was working before started
to show compilation errors. The latest DaCe is moving a data access
to an inter-state edge. For the data-access, the symbols that define
array strides are needed for code generation. The SDFG was validated,
before and after the simplify pass, but it did not compile for CPU.
When skipping the simplify pass, the compilation did work.
The problem has been narrowed down to the scalar-to-symbol promotion,
which is moving a data access to an inter-state edge. Then, the
method _used_symbols_internal needs to be update to account for data
containers, including symbolic shape and strides.
This commit contains a unit test to reproduce the issue and verify
the proposed fix.