New way to call a `CompiledSDFG` object.
Before `__call__()` has to do some reordering (from the one given by `argnames` to `_sig`) and transforms the `ndarray`s to pointers, which was a quite expensive task.
This commit introduced the `_fast_call()` method, which allows to bypass these operations.
It expects that the arguments (the one for the call and the one for the initialization) were prepared outside, and no further check is done.
This can be used for example during benchmarking to avoid measuring the overhead of the previously mentioned transformations.