Updated and fixed the MapDimShuffle tranformation.
The reordering of the dimension was quite strange, it is now much more cleaner.
Furthermore, the `tile_sizes` parameter of the map range was ignored, i.e. not changed.
Furthermore, the parameters are now correclty copied.
Before it was just assigned, thus if later the parameter list of the map was again changed this effect would propagate to _all_ maps that where treated.