`DeadDataFlowElimination` will add type hint when removing a connector (#1499)
Issue #1150 that `DeadDataflowElimination` removes a connector from a
Tasklet which leaves a variable without type hint.
This PR tries to fix this bug by adding a type hint expression for a
variable which is used in the tasklet. It adds the type hint only if the
variable is used inside the tasklet code (I checked using
`ASTFindReplace`).
The PR also adds a test which is literaly the code presented in #1150
and asserts the presence of the type hint and checks if it compiles.

### May need confirmation
- [ ] Did I use `ASTFindReplace` correctly?
- [ ] If the type inference fails no type hint is added. Is it the right
solution?
- [ ] Does the test even make sense? (I don't have much experience in
unit testing).

This is my first PR for this project so be patient with me.

---------

Co-authored-by: alexnick83 <31545860+alexnick83@users.noreply.github.com>