Generalize StructArrays to ContainerArrays and refactor View class structure (#1504)
This PR enables the use of an array data descriptor that contains a
nested data descriptor (e.g., ContainerArray of Arrays). Its contents
can then be viewed normally with View or StructureView.
With this, concepts such as jagged arrays are natively supported in DaCe
(see test for example).
Also adds support for using ctypes pointers and arrays as arguments to
SDFGs.

This PR also refactors the notion of views to a View interface, and
provides views to arrays, structures, and container arrays. It also adds
a syntactic-sugar/helper API to define a view of an existing data
descriptor.

---------

Co-authored-by: Alexandros Nikolaos Ziogas <alexandros.ziogas@inf.ethz.ch>