Fixed a bug in the map fusion transformation.
Essentially the bug happens because the transformation considers any array with one elements as scalar, including arrays such as `(1, 1)`.
This commit allows the scalar branch of the transformation to handle this kind of situation.
Another solution would be to "promote" it to an array.

In addition this commit adds some test cases for these kind of events.