Updated and fixed the MapExpansion transformation. (#1532)
Before the transformation ignored the tiling parameter of the range,
this was fixed. The transformation is now also able to limit the
expansion. Thus it is possible to only expand the first k dimensions,
the remaining dimensions will be kept in a multi dimensional map.

This is a feature that will be need in GT4Py.