Adds support for ArrayView to the Python Frontend (#1565)
The refactoring of Views in PR #1504 led to the creation of the
ArrayView type. This PR addresses an issue in the Python ProgramVisitor,
where ArrayViews are not recognized properly as Views (of Arrays),
leading to a NotImplementedError. The fix is simple: when checking if a
container is an Array or a View (of an Array), instead of making a
direct equality comparison to Array or View, a subclass comparison
against Array is performed. The latter returns true if the container is
an Array or any Array subclass, including ArrayViews.