Control Flow Block Constraints (#1476)
This PR implements pre- and post-conditions on control flow blocks, as
well as invariants. This is a feature that was discussed in the last
DaCe Workshop of 2023 but has not been implemented yet since then.

These invariants serve as helpers in analysis of the SDFG and may in the
future be used to add runtime checks / assertions (optionally). A pass
can be used to attempt automatic derivation of such constraints, or they
can be manually set as properties.

This PR adds the scaffolding for this and for now implements a single
auto-constraint-derivation criterium, which states that parameters used
to determine data container sizes are always `>= 0` and optionally
assumed to always be `<= MAX_N`, where `MAX_N` is a configurable
analysis pass parameter.