fixed an error regarding the `pow()` function of the DaCe runtime.
Currently there was an inplementation for `T pow(T, T)`, where `T` is a template and two for the case where `T` is a signed or unsigned integer.
This is a problem if we consider the call `pow(double, long long)` because the call is ambigious and results in a compiler error.
As a solution I added the function `T pow(T, E)`, where both `T` and `E` are templates.

I also think that the implementation of the int case could be improved because currently it is a simple for loop.