Scalar return values are now invalid (#1609)
Due to how the code generator works a scalar return value is not
possible and it does not generate an error.
Returning a scalar is a desired feature as there is code in
`CompiledSDFG` for that, but it does not work.
The frontend works around this by simply promoting everything to an one
dimensional array.

This commit introduces:
- A change to validate to reject such SDFGs.
- A test that can test the feature, but is disabled, since it does not
work.
- A test for the new restriction.