Updated the transformations.
In the previous versions I tried to simply ignore the reshaping Memlet case, however, this break other tests.
In this version reshaping Memlets are not turned into a view, i.e. the source array is not removed but turned into a view.
This fixed the other tests again.

It is not the best solution but it works.