Allow to parameterize the `#pragma unroll` pragma.
Before it was only possible to instruct DaCe to add this pragma or not, but it was not possible to specify a value.
This commit adds a new property to the `Map` that can be assigned to a value that should be added to the pragma.
TThe main intention of this is to prevent unrolling altogther, by be able to issue `#pragma unroll 1`.