Implement user regions and function call regions (#1623)
This is part of my GSoC project.
This PR implements 
- user defined control flow region using the new `dace.named` context
manager. The graph corresponding to that code section is wrapped the new
`NamedRegion`, a subclass of `ControlFlowRegion`.
- function call regions, regions which represents a function call in the
source code. The function has to exist as a global SDFG or
`@dace.program`.

### Needs confirmation

- I changed
[tests/python_frontend/multiple_nested_sdfgs_test.py](https://github.com/spcl/dace/pull/1623/commits/d693269dded722ef2d49b54a0167cededbe99a36#diff-829b32045e5b9c0371308f9dae0bbe7c04fa1d80b6273ec7ba2ffda9a8c83d5f)
and use regex on state's label for finding the right state to be tested.

---------

Co-authored-by: Philipp Schaad <schaad.phil@gmail.com>