More towards bug compatibility.
Be a bit more restrictive what a shared intermediate is.
If an intermediate node is used somewehere else (or rather its data) then the transformation will not fuse it.
Note that this is not a bug in the MapFusion transformation, but in the auto optimizer (most likely SubgraphFusion`).
This commit essentially adds a feature to disable some improvements of the transformation such that auto optimizer works.

To see this use `tests/npbench/polybench/correlation_test.py`.
Use the MapFusion alone and you will see it works.
Then use auto optimization and you will see that it does not work.