Updated the behaviour of `Subset.covers` and related functions.
Before all cover functions ignored the case if the subsets have different lengths.
The effect was that the subsets `A = Range.from_string("i, k, 0:N, 1")` and `B = Range.from_string("i, k")` would cover each other.
This case was considered undefined behaviour.

This commit changes the behaviour of these function such that for covering it is required that the subsets must have the same dimensions.