Refined my fix in the `_read_and_write_set()` function.
I realized that an undefined dst_subset of an incomming edge of an access node is only possible if the access node is a scalar.
So I modified the function such that the corrct subset in that case is generated.

This case happens if a single entry of an array is read and written into a scalar.
Then only the source subset must be defined.