Fixed the bug in `_read_and_write_sets()` that made `tests/numpy/ufunc_support_test.py::test_ufunc_add_accumulate_simple` fail.
However, I am pretty sure that this is a compability bug.
The function essentially removes accesses from the read set if there is a matching access in the read set.
There is nothing wrong with that, however the original code only filtered that out when the data of the in- and output memlets refered to the same data.
However, this is wrong, because there is no guarantee that this holds as the memlet's data member must not coincide with the node under consideration, but can be another (that is connected).

But if thsi check is removed then `tests/numpy/ufunc_support_test.py::test_ufunc_add_accumulate_simple` will fail.
So I put it back.