Modified how the optional stuff in the properties works.
Before there were `optional` and `optional_condition` (which had the oposite meaning, i.e. not optional if the callable evaluates to `True`).
This was all replaced by the `serialize_if` callable, which if it returns `True` what the default does the property is serialized and if it rerturns `False` the property is not serialized.
This commit also introduces the `dtypes.CPU_SCHEDULES` collection.