Updated the trivial map elimination transformation.
It is now able to cleanly handle the case of dynamic map ranges.
In that case the map is no longer removed, but one map parameter is added back, whcih is needed for syntactical reasons.

Furthermore, the transformation will now ignore GPU maps, this is because they may be needed for purly syntactical reasons.