Modified the `SetProperty`.
Because of the implementation of the `__set__()` and `__get__()` functions such properties modeled an immutable set, by essentilly copying the value that was assigned to it and returning a copy of the internal set.
However, because `__get__()` returned a regiular `set` this object was mutable, but all mutation did not affect the property, which is a rather unintuitive behaviour.

The effect is, that every single `assert` in the code bellow fails.
```
import dace

@dace.properties.make_properties
class Test:
    member = dace.properties.SetProperty(element_type=int)

    def __init__(self):
        self.member = set()

a = Test()

assert a.member is a.member
a.member.add(10)
assert len(a.member) == 1
assert 10 in a.member
```

The new implementation now stores a `frozenset` internally, thus the first assert will now pass.
Furthermore the line `a.member.add(10)` will now fail, since a `frozenset` can not be modified.

For backward compability the property pretends to be a regular set.