Updated the `replace_propertiers_dict()` function.
Before the function updated the underlying data memember (`_ignored_symbols`) of the `ignored_symbols` `SetProperty` of the `Tasklet` directly.
However, since this function has been turned into a frozenset this no longer works.

As far as I can tell this is the only case where this happens.
So I decided to perform the full updating, less efficent but it only happens for C++ Tasklets, so it should not be a big issue.
In my view the accidental update of `SetProperties` (the behaviour is not documented) is much more important.