Made the `SDFGState.add_mapped_tasklet()` more convenient.
Before if the user wanted to supply in and output nodes he had to present a `dict` that maps the data name to the access node.
However, because of the rules of a valid SDFG the key of that `dict` was always the same as the data the access node this information is redundant.
Thus this commit allows to only pass the access nodes.