Updated `PruneConnectors` again.
After a discussion with Alex Nik, I removed the check for symbols.
This is because it is impossible (validation of nested SDFGs) to pass them through connectors.
Instead symbol_mapping should be used, thus we do not have to check for them.
An SDFG that does this is invalid anyway.