The `PruneConnectors` now again has the old behaviour in when an input is maintained and when not.
However, this has some consequences.
I added a test, that shows that this leads, depending on the memlet configuration to different outcome of the transformation.
Okay, it is only affects if the transformation can be applied or not, but still.
This is also in line with my [issue#1643](https://github.com/spcl/dace/issues/1643) that shows that this is a problem.