Found the offending code.
Because it was originally inside a while loop the `find_new_name()` function was actually called multiple times.
First with the "dirty" name, that still contained dots, which was then replaced.
Because the first time the `add_datadesc()` function was called with that name and `find_new_name=True`, the cleaned name was already known.
Thus it would again be passed to `_find_new_name()` but this time the name was cleaned and augmented (because it once passed through `_find_new_name()`).
The code introduces exactly replicated this behviour.
If it should be disabled, then the cleaning and the first call to `_find_new_name()` can be swapped and the if removed.

Note this also means that the original `name` contained dots _and_ was added to `_arrays`.
However, it did not trigger an error in validation, so it is probably fine.