Patched the Python interpreter to properly handle process grid and such.
Before the MPI replacement functions created pseudo-scalars with the same name as the corresponding entity, i.e. if there is a process grid with name `__pgrid_1` then there was a scalar with that name.
According to the comment, this scalar is there to prevenbt the removal of some dummy nodes.
However, as a side effect the Python frontend was also finding them and had not to wory about the MPI types, at least in assignment statements.
But because we disallowed that different entities had different names (except for constants) this no longer worked and the frontend had to explicitly handle them.
This commit added the facilities for this.