Fix array indirection to memlet subset promotion (#1406)
The current solution is rather hacky. I want to run the tests first to
see the impacts of this change.
Additionally, there is no test yet, because validation doesn't catch the
erroneous SDFG yet.

Overall, it's not clear currently how to solve the issue and the PR
might change as we progress...