Modified the `CopyToMap` transformation, with regard to strides.
The original version of the transformation tested if the strides were the same and if so, declined to work.
I do not fully understand why this behaviour is there, but this commit renables this behaviour for teh sake of compatibility.
However, it adds the possibility to disable this, whcih is done inside bypass tests, which is a very stupid name.