Control Flow Raising (#1657)
This PR mainly provides control flow raising passes for the new
intrinsic control flow constructs (Branches and loops) in SDFGs. In
addition to raising, the state and control flow reachability passes have
been adjusted to faithfully work with the intrinsic control flow
constructs.

Along with the raising and reachability passes, a few important bugfixes
and a general cleanup is included in the PR, but no other functionality
is changed.