Better `CopyToMap` (#1675)
By default the transformation uses some linearization followed by a
delinearization approach, while this is needed to copy certain shapes,
it is unnecessarily complicated for memlets such as `a[0:10, 20:30] ->
40:50, 60:70`.
This PR adds special cases where the source and destination subset has
the same size and transforms it to a simple copy.
It also supports the case where some dimensions are one, i.e. memlets
such as `a[0:10, 0:10] -> 0:10, 1, 0:20`.

For all cases tests were added.
Most importantly the transformation now applies if the strides are the
same.
Before this case was blocked.

This PR helps to _avoid_ errors that are related to
[Issue#1674](https://github.com/spcl/dace/issues/1674), but it is not a
fix or a solution to it.