Add a `SubrangeMapper` helper class.
Its functionality
===
Equipped with a `src` and a `dst` range of equal volumes but possibly different shapes or even dimensions (i.e.,
there is an 1-to-1 correspondence between the elements of `src` and `dst`), maps a subrange of `src` to its
counterpart in `dst`, if possible.

Note that such subrange-to-subrange mapping may not always exist.