Cherry-picked a handful of intrinsic related commits out of `multi_sdfg` branch. (#1728)
These commits:
1. are chronologically early
2. fairly independent and touch a small number of files
3. thematically fit together.

This is a follow-up on the discussion about incrementally moving the
stable features from `multi_sdfg` branch into `main`. All of the commits
are cherry-picked to preserve the history and to have easy
fast-forwarding of the `multi_sdfg` branch.

Some commits had to be slightly modified to resolve the merge conflicts
during cherry-picking, but otherwise they are left as-is.

---------

Co-authored-by: Marcin Copik <mcopik@gmail.com>