Fixing `SDFGState._read_and_write_sets()` (#1747)
Removed some asserts that were to strict from
`SDFGState._read_and_write_sets()`.

If a subset is `None` the function assumes that corresponding array is
fully accessed.
Before, there was an `assert` that ensured that this assumption is true.
However, the implementation also assumed that this could be verified,
either because the same symbols were used or the size was known.

But this if not necessarily the case, if the two involved arrays uses
different symbols, that however, are always the same.
The other reason was that before the size was estimated using
`total_size`, which has a well defined meaning for an `Array` but in
case of a `View` its meaning is not fully clear.

For these reasons the asserts was removed.

---------

Co-authored-by: Tal Ben-Nun <tbennun@users.noreply.github.com>