Updated the strict dataflow mode.
It is now made the default.
Also added some tests to check if it is properly working.

This mostly refactors the implementation.
It should be the same as before, however, the only thing that I really removed, that could become a problem, is the check if it is used in the same state.
This was always a hack, and the check should have been that it is not used in the same data flow.
But I think it was redundant anyway.