Added a new check to the map fusion in case of shared intermediates.
It essentially check if the intermdeiate, that should be turned into a shared intermediate, i.e. a sink node to the map exit, is used in the data flow downstream.
This is needed because some DaCe transformations do not correctkly check for the existence, it even seems that it is killed.