Performance improvement in the `add_{in, out}_connector()` functions of the node.
Before every time the function was (shallow) copying the `dict` upon a change.
This was done to make use of the implicit casting functionality of the `DictProperty`, however, it is not very efficient.
This commit changes the implementation of the function to perform the casting explicitly, thus avoiding the need to copy the `dict` every time.