The test `test_fusion_with_nested_sdfg_0` is now explicitly constructured.
Before the output edges were before set to dynamic.
However, this was not true as it was always set, thus the new map fusion did not fuse them.
My first attempt was to just disable the `dynamic` property, but now the SDFG is generated manually.
It is almost the same, but uses lesss symbol, as it was simpler to implement it this way, and we are now using float.