Fixed the problem in the heat test.
The issue was similar as before.
When I computed the name of the intermediate transient then I used `sdfg.node_id(state)` to get the state ID.
However, now if the state is part of these recursive control flow regions then this may not work, because the state is not a direct node of the SDFG.
However, if I use `self.state_id` then it works, this is what the old MapFusion was doing.