Found a case that was not handled.
The transformation checks if the first map satisifes the data dependencies of the second map.
For this is looks at the writes and reads of the intermediate.
It also checks if, a data container is used as input of the first and as output of the second map, if the access is pointwise and can be fused.

Furthermore, it was allowed that the intermediate is also used as input to the first map.
However, in that particular case, it was not checked if the the reads and writes of the first map alone to the intermediate are valid.
I.e. it could read read `A[i]` but write `A[i+1]` which would cause problems (note that this usage is botherline legal anyway.
This commit adds a check to make sure that this is not the case by enforcing if a data container is used as input and output of the first map and also as intermediate node then its read must be pointwise.
Note that if it is not an intermediate node, i.e. not also read by the second map, then this rule does not apply.

NOTE: It is forbidden that the intermediate is used as intermediate and output of the second map.