Fixed a bug.
I was using the `symetric_difference()` instead of the `difference()` function to handle the inter state edges.
However, this also lead to a clarification and a stricter handling of interstate edges.
I.e. as soon as data is accessed on an interstate edge, it can never be exclusive.
Furthermore, now not only the read symbols but all free symbols are used.