Made the `MapFusion` transformation stateless.
Instead of scanning the SDFG once it will now scan the SDFG for every intermediate that is processed.
However, there is a fallback that allows to use the data that is computed by `FindExclusiveData`, which is not yet pressent, but will be introduced by [PR#1906](https://github.com/spcl/dace/pull/1906).
But there are a few things to consider:
- The `FindExclusiveData` is not passed as argument but through the `self._pipeline_results` variable.
	This is done to make it stateless, as it would be the same as before, just computed outside.
- Before the cache stored the names of the data that could be bot be removed, now it is the data that can be removed.
	This makes it a little bit more stable.
- Even if we use the `FindExclusiveData` mechanism, we also have to perform some limited checks.
	This is because the old scan computed some slightly more restrictive.

However, the mechanism is not yet introduced.