Fixed a bug.
The problem was that with the new hirarchical structure, the previous version does not work.
It generates an error that some state, the one inside the loop region is not known.
I have no idea why it does not work, but I accept it.

This caused that some fusions did not happen.
What was interessting that then the composite fusion generated a wrong SDFG.
I am not sure what the problem is exactly.

But again `match_exception` was my friend.
It should be the default anyway.