Added a warning to `FullMapFusion` pass.
This is because the `pipeline_results` is only pressent during `apply()` and not during `can_be_applied()`.
I do not want to add the result at construction to `MapFusion` because this would make the transformation stateful.
Somebody mentioned at the meeting (acceptance of the MapFusion transformation 2025-01-22) that this should not be.