Found a way to work around the limitations of the `_pipeline_results`.
I overcome the limitation by adding my own version of it, which is called `_single_use_data`.
It is explicitly set by the `FullMapFusion` transformation and later reset, i.e. set to `None`.

It is important that there is no constructor argument to set this.
This is a deliberate choice to make the transformation as stateless as possible.
Because, this was a remark made in the DaCe meeting that the transformation should be stateless, so I did it.
Since the goal to make it stateless was the whole reason to make this separation in the first place.