Fix read_and_write_sets() for ControlFlowRegion (#1920)
The `PruneConnectors` transformation relies on `read_and_write_sets()`
to identify connectors on a nested SDFG that refer to unused data
containers. With the introduction of `ControlFlowRegion` nodes, the data
containers can now be accessed by symbolic expressions used as
conditions in such nodes. This case was not considered in baseline.

---------

Co-authored-by: Philip Mueller <philip.mueller@cscs.ch>