Before we will enable every test manually, we will first revert the changes in auto opt.
There is no real change, it is just that now another component is disabled and we will now see what happens.
If this does nto work, we are forced to essentially every combination of tests using the CI.
Does this sounds like fun.

Revert "`auto_optimizer()` now uses the new `FullMapFusion` transformation."

This reverts commit 0f5228bca7194a4da79fa0d6ebd91fb509d1262c.