It is very interessting.
When I run with all single test files then it fails, when I run with one part it (the one not listed in `SMALLER_FRACTION`) then it works.
And if I only add them that are listed in `SMALLER_FRACTION` it still fails.
So lets see if it is still fails if all are run.