Fix tutorial codegen (#1925)
The tutorial `codegen.ipynb` has a small bug, namely the
`ControlFlowRegion` is not in the correct namespace (as used in the
tutorial).
This PR fixes the issue by importing `ControlFlowRegion` in the `dace`
namespace.

Additionally I added a test for the notebook tutorials to make sure they
are not broken.
At the moment I added only the "getting started" and "codegen"
notebooks.
These tests require new dependencies `ipykernel` and `nbconvert`.

The above may or may not be the preferred solutions!

---------

Co-authored-by: Tal Ben-Nun <tbennun@users.noreply.github.com>