Moved the test for negative sized subsets from the Memlet to the `vaidate_state()` function.
The reason is that in some cases this is valid, for example if an edge connects an AccessNode and a MapEntry, because, in that case the map might not be executed.
Since the Memlet does not have access to its source and destination node it can not check that, so the test was moved to a location that can do this check.
However, it only does the check for AN to AN connections, which is a bit restrictive, but this is something for later.