Updated MultistateInlining (#1980)
This PR addresses
[issue#1959](https://github.com/spcl/dace/issues/1959), the issue was
about that the inlining transformation destroys the SSA invariant we
maintain inside GT4Py.
This invariant states that in every code path there is exactly one
AccessNode with that is used to write to a data container.

The PR solves this issue by modifying how the nested SDFG is isolated,
before this was done in two steps and is now performed in a single step.
The main idea is that the number of writes is preserved.