Updated `MapFusion` (#1993)
This PR adds two new functionalities to `MapFusion`.
Until now it was only possible to fuse two Maps if the intermediate node
has a single producer (the top Map).
But this is not strictly required, it can have as multiple producer, the
important thing is that the second Map only consumes data that have been
written by the first Map.

The second functionality, is actually a fix.
Consider the following situation:
```python
for i in range(10):
    tmp[i] = a[i] + 10
for i in range(10):
    b[i] = tmp[i] * a[i]
```
Both the first and second Map have an input edge from an AccessNode
referring to `a`.
There was a functionality that was supposed to remove the edge going to
the second map, i.e. reuse the edge that was going to the first map.
However, this was not properly working and this is fixed.
Note that it is not done for the outputs, i.e. if both Map would write
to the same data then the new combined Map would have to outgoing edges
to that data.
The reason for this is because it would need a special case,
furthermore, this is a rare case.
Furthermore, there was an error regarding the handling of nested Maps.
A test for that will be added in the [parallel Map fusion
PR](https://github.com/spcl/dace/pull/1987).

---------

Co-authored-by: alexnick83 <31545860+alexnick83@users.noreply.github.com>