Fixed a bug.
When the SDFG was restored from file, then the list of positional arguments was not set, because the argument has a default value, `None` which means "no positional arguments", the restored SDFG would not accpet any positional arguments, althought the version created by `sdfg.compiled()` would accept them.
This Commit also adds a warning in the constructor, if no positional arguments where passed, to the constructor of `CompiledSDFG`, but the SDFG has arguments.