The code generator now uses the helper function to turning a Memlet into a Map.
The reason for this is quite simple.
Before the code generator used `CopyToMap` to turn unsupported Memlets into Maps.
However, that transformation operated on AccessNodes, but if there are multiple edges between the nodes then it was unspecific whcih one was turned.
In the worst case the one that would be supported was turned into a Map but the unsupported not.
Since the `memlet_to_map()` function operates on edges this issue is no longer possible.