Created a new state fission function.
It is designed in a similar way than `isolate_nested_sdfg()` function, i.e. it preserves the number of writes to nodes.
The change here is, that it might happen that some consumer of `subgraph` end up in the first state.
However, The old version also had that problem, because I could pass a simple Tasklet as subset and then one would end up with an invalid function.