F2dace/func type fix (#2026)
Fix type inference for functions that are part of an expression where
the return type cannot be deduced trivially.

For `foo() + 3`, type inference works even if the type of foo() is
unknown or `void` because the hierarchically "maximum" type is
automatically deduced.

However, type inference fails for expressions like `foo() ** 3` if the
type of `foo()` is unknown because the `**` operation is handled by the
intrinsic function `__POW_()`. Thus the type of `foo()` must be known at
this point.

The fix consists of two parts: First, the type is attached to the
function declaration in `ast_components.py::function_stmt`. Second, the
type is retrieved from a list of function declarations in
`ast_transforms.py::Call_Extractor`, which transforms each function call
into a subroutine call+temporary variable.

Co-authored-by: Faveo Hörold <faveo.hoerold@cscs.ch>