Only report cycles if we found some (#2055)
## Description

The issue here is that `find_cycles()` returns a generator, which is
truthy even if there are no cycles. The proposed solution is to evaluate
the generator. That way, if the generated list is empty, it will be
falsy and we won't report cycles if there are none. This allows
developers to see the other error message, which is currently
unreachable.

Co-authored-by: Roman Cattaneo <1116746+romanc@users.noreply.github.com>