Avoid losing symbols in `traverse_sdfg_with_defined_symbols()` (#2061)
From what I can tell, `traverse_sdfg_with_defined_symbols()` is supposed
to traverse the SDFG, accumulating symbols as it runs down the tree. The
current implementation resets the dict of "already known symbols" when
recursing down into a state that is the source of an edge. That seems
odd. In particular, not even the globally available `sdfg.symbols` are
marked as defined anymore with that reset.

The proposed fix passes along already known symbols like in all other
cases, ensuring that the dict of known symbols only ever grows.

This is a port from PR https://github.com/spcl/dace/pull/2054 to the
mainline branch.

---------

Co-authored-by: Roman Cattaneo <1116746+romanc@users.noreply.github.com>