Remove Legacy Control Flow Detection and Improve ControlFlowRaising (#2045)
This removes the legacy control flow traversal detection methods from
DaCe. These methods were used in two key places:
1. Codegen, to detect and determine what kind of control flow to
generate
2. Schedule tree creation (SDFG to STree), again for control flow
detection.

With this PR, a few key bugs in `ControlFlowRaising` were addressed,
`ControlFlowRaising` now quarantines unstructured control flow
conservatively (based on the smallest single-entry-single-exit region
surrounding edges of unstructured control flow), and all control flow
related questions were relegated to the SDFG and its control flow
regions. The old option to automatically detect control flow has been
kept, and instead of the legacy control flow detection, the
`ControlFlowRaising` pass is now run in codegen to satisfy that setting.
I suggest that the default of that setting be changed to `False` in the
future of DaCe 2.0 before it is phased out entirely, since the goal is
to have control flow regions explicitly already at that point, and
performing the check redundantly is an additional performance penalty
that is not needed.

One note on the functionality of control flow detection w.r.t. to this
change: Legacy control flow detection tried to generate switch-case
statements where possible. The current lowering of `ConditionalBlocks`
to code does not yet support the generation of switch-case statements. I
suggest a feature request issue is added to DaCe 2.0 to add this, but it
needs to be made clear how this behaves (it is different than a
traditional switch-case in the sense that no fall-through is allowed -
which was already the case with how it was implemented in legacy control
flow detection). This was quite opaque before.