Fixed unsafe modification of `ScopeTree.children` (#2066)
The `scope_tree_recursive()` function used an unsafe way to operate on
the return value of `SDFGState.scope_tree()`, it was modified.
The main issue with this was that the result of the `scope_tree()`
function is cached.
Thus subsequent calls, will get see the modified versions of the
`ScopeTree` entities.
This PR solves the issue by:
- Adding asserts to ensure that no "sharing" is happening.
- Surrounds the calls to `scope_tree()` by calls to
`SDFGState._clear_scopedict_cache()` to ensure that a clean versions of
the `ScopeTree` is obtained and the modified versions are not used later
on.

Furthermore, the function also added a `.state` attribute to the
`ScopeTree`, which denoted the original state it was found in.
This was not documented (also some other part of the code was depending
on it) and a description was added.

In addition the `gpu_map_has_explicit_threadblocks()` function was
optimized.
Before it traversed the whole graph twice now only once.