Updated teh `SubgraphView` mainly the `edges()` function.
Before the edges were stored inside a `list` thus every chekc, i.e. `in` resulted in a linear time operation.
This was particular devastating for `edges()` since it degraded to quadratic runtime.
This PR changes the undelrying data structure to a `dict` which is still ordered but allow much faster `in` tests, which is the main workflow.
The caviat now is that `nodes()` result in the construction of a `list`.

This PR does not change the behaviour or the observable behaviour of the class, it just makes it faster. The only change is

Note:
I would propose to store the list of nodes inside a `set` to begin with.
The only reason why they are now in a `dict` is because they were originally ordered by their ID in the parent graph.
Since the nodes are not necessaraly consequitive, I do not really see why?