Modified How State Fission Works (#2047)
This PR changes how state fission is performed within DaCe.
Before, there were two functions for it `state_fission()` and
`state_fission_after()` (which was used only in two places and did not
have a unit test).

As a smaller change, this PR removes the `state_fission_after()` as it
was used at only one place, had no description and no unit test and was
replaced with `state_fission()`.

Furthermore, this PR changes how `state_fission()` works, the redesign
was inspired by the change introduced by
[PR#1981](https://github.com/spcl/dace/pull/1981) which adds
`isolate_nested_sdfg()`, which ensures that the number of "writing
nodes", i.e. AccessNodes with incoming connections is preserved.
The changes of introduced by this function goes into the same direction
as `state_fission()` will now also conserve this property.
This has some behavioural change on the function, but also fixes some
behavioral problems.

As argument `state_fission()` takes `subgraph`, which is a list of nodes
that should be moved to a state that is sequenced before the state they
are currently located in.
The function will now automatically add all dependencies of these nodes
to the list of nodes that have to be relocated, currently this had to be
done by the caller.
The other and bigger change is, that now also nodes that are downstream
of the passed `subgraph` might be moved to the first state as well.
There are two reasons for this, the first is to ensure that the number
of "writing nodes" does not increase.
The second reason is that the cut, i.e. which nodes can go to the first
node, is not arbitrary and only Memlets whose source is a non view
AccessNode or empty Memlets can be cut.
Previously, it was the callers responsibility to ensure that, now the
function does this.

Another reason why nodes are moved to the first state is scope.
This is because a split can only be done on the top level, for example a
Map can not be distributed over the two state (for example the MapEntry
and some nodes are placed in the first state, while the rest, including
the MapExit ends up in the second state).
Thus the function will now make sure that this does not happens.
For this it will make sure that if a node is not at the global scope of
the state, that its surrounding scope is also moved to the first state;
This happens recursively to ensure that nested scopes are handled
properly.

---------

Co-authored-by: Philipp Schaad <schaad.phil@gmail.com>