Better Use Of Simplify (#2093)
Simplifying expressions is a relatively expensive operation.
As a performance improvement the code was refactored in such a way to
that calling simplify is avoid or deferred as much as possible.
The general idea was to first try doing it without simplify and then if
no "positive" answer was obtained try it again with simplify.

The code refactoring mostly preserves the previous behaviour, with the
exception that some `TypeError`s, which are generated if the truth value
of SymPy expression was determined were removed, since they were never
propagated to the user.
Instead they now stay within the function.

It was also discovered that the assumption of `symbol_positivity` was
used in a very aggressive way.
It is even possible that it is used in places where it should not.
However, in the 2025-07-24 DaCe meeting it was decided to leave the
current behaviour, since major changes are expected soon in that part of
the code.