Avoided costly copy operation (#2096)
During a profiling it was noticed that a huge fraction of the time was
spend in `__deepcopy__` of Memlets.
After some investigation it turned out that this was caused by copying
SymPy expressions.

In SymPy symbols are
[immutable](https://docs.sympy.org/latest/explanation/glossary.html#term-Immutable)
the documentation makes this quite clear.
Thus copying such an object should be trivial, however, for some reason
it is not (apparently SymPy does not define the `__deepcopy__()`
function).

Thus this PR adds a `__deepcopy__()` method to the:
- `Memlet`
- `Range`
- `Indices`
- `DebugInfo`

This method only copies the members that are mutable but does not touch
the immutable parts.
The effect is a speedup of about >10%, the side effect, however, is that
`deepcopy()`, at least for some object, is no longer deep, however, due
to the immutability, there is no behavioral change.

Furthermore, this PR fixes the `__add__()` method of the `Range`.