Fix codegen for cuda memory pool (#2103)
This PR addresses a bug in CUDA code generation for the `Array.pool`
feature. This feature allows to allocate global arrays on the memory
pool associated to the stream.

Considering the SDFG in the newly added test case, the main baseline was
producing the following CUDA code:
```
void __program_test_memory_pool_if_states_internal(test_memory_pool_if_states_state_t*__state, double * __restrict__ A, int cnd) {
    if ((cnd == 0)) {
            double * __tmp0;
            DACE_GPU_CHECK(cudaMallocAsync((void**)&__tmp0, 10 * sizeof(double), __state->gpu_context->streams[0]));
            __dace_runkernel_write_zero_map_2_1_1(__state, __tmp0);
            DACE_GPU_CHECK(cudaMemcpyAsync(A, __tmp0, 10 * sizeof(double), cudaMemcpyDeviceToDevice, __state->gpu_context->streams[0]));
            DACE_GPU_CHECK(cudaStreamSynchronize(__state->gpu_context->streams[0]));
    } else {
            __dace_runkernel_write_cond_map_3_1_0(__state, A, cnd);
            DACE_GPU_CHECK(cudaStreamSynchronize(__state->gpu_context->streams[0]));
    }
    DACE_GPU_CHECK(cudaFree(__tmp0));
}
```

The array pointer `__tmp0` is out of scope when `cudaFree` is called.
The origin of the problem was that arrays with `Scope` lifetime were
treated as global arrays. Besides, `cudaFree` was called instead of
`cudaFreeAsync`.