Conditional Map Interchange (#2097)
This PR adds a new ConditionalMapInterchange transformation accompanied
by tests. The transformation looks for maps inside conditional blocks
such as:
```py
if cond:
  for i in map[0:10]:
    A[i] += 1
```

and interchanges the conditional block with the map:
```py
for i in map[0:10]:
  if cond:
    A[i] += 1
```

This is particularly useful if there is an outer map, which can then be
fused with the inner map thanks to the interchange, increasing the
number of parallel threads.

---------

Co-authored-by: Philipp Schaad <schaad.phil@gmail.com>