Fix the bad `zip()` calls in redundant array transform (which doesn't check for length equality). (#2105)
This fixes the problem of incorrect compressed memlet calculation, where
the memlet chain going through a transient does not carry the same
volume everywhere.

This used to result in an OoB memlet in the test case, where that
transient was turned into an incorrect view, which was then removed
leaving an OoB memlet. The solutions seems to be: there was an incorrect
`zip()` in `can_be_applied()` that does not make sure the two shapes are
really matching. Just handling that gets us out of the OoB memlet case.

---------

Co-authored-by: Pratyai Mazumder <pmz@pmzx.local>
Co-authored-by: Pratyai Mazumder <pmz@student-net-ox-dock-1-a-023.intern.ethz.ch>