Fix: Memlets in StripMining transformation (#2118)
**Bug Description:**

To put it simple, the `StripMining` transformation essentially performs
tiling- that is, it adds a **new outer** map which encloses
an existing one, where the new map runs over chunks (=tiles) of the
original iteration space, while the inner map handles
individual elements inside each chunk. The inner map is the original
map, where the range just go adapted.

Now, the new (outer) map needs to be connected to the original one, and
the dataflow between them must be captured correctly.
This essentially involves analyzing which parts of a data descriptor
should flow for each chunk. For the sake of simplicity, one
can think of this process as: copying an existing memlet (flowing to and
from the original inner map) and update it to create
the correct memlets between the two maps. The screenshots below help
visualize this - you can clearly see the similarity
between the two highlighted memlets.

However, this copying step is problematic when an `other_subset` is
present - in such cases, it should be explicitly set to `None` for
the newly created memlet. Otherwise there can be a dimensionality
mismatch between the two subsets leading to an invalid SDFG.

**Fix Description**

Set the `other_subset` to `None` for memlets created between the two
maps. Additionally, a test case has been added which now passes and
would have failed before this fix.


**Failing Example**

After the fix, ```[0] -> A[block_i : Min(63, block_i + 31) + 1]``` in
the right image will become
```A[block_i : Min(63, block_i + 31) + 1]```.

<p float="left">
  <img width="45%" src="https://github.com/user-attachments/assets/33c14e92-43c6-4d93-a77e-87691fa8f9b8" />
  <img width="45%" src="https://github.com/user-attachments/assets/76b7f030-91c6-424d-a91b-3ba9df710840" />
</p>