Fixed `MapFusionVertical` with NestedSDFG (#2127)
`MapFusionVertical` merges two consecutive Maps.
In certain cases the intermediate between the two Maps is replaced by a
"reduced version", usually a scalar but also smaller arrays, in most
cases of constant size.
Usually this is not a problem except in cases when the strides of the
original intermediate are used, for example if it is mapped inside a
nested SDFG.
Thus if we replace the intermediate, we also have to update the strides
inside the nested SDFG.

The implementation introduced by this PR has some limitations.
The most important one is, that the update is not recursively.
Thus, if the data container inside the nested SDFG is passed into
another nested SDFG the transformation will simply not apply (which is
an improvement, since before it applied and did not updated).