fix: support de-aliasing SDFGs with in/out conns of the same name
This commit adds support for de-aliasing nested SFDGs with input/output
connectors that have the same name. When unsqueezing memlets (in the
middle of the function) it's important to pick the right "outer memlet"
and if we don't separate inputs and outputs, we (randomly) choose either
one for both, which can lead to subtle bugs.