used_symbols() call does not match argument list (#2128)
**Description**

The `used_symbols(...)` method is called incorrectly in the
`used_symbols_within_scope(...)` function.
From the `nodes.py` file, we can see that only the `NestedSDFG` class
implements this method. However, it is currently invoked with two
arguments instead of the single argument it expects. So we need to call
it for `NestedSDFG` instances only with
one argument.

In addition, the `hasattr(n, "used_symbols")` check can be removed and
replaced with a direct `isinstance` check for `NestedSDFG`, extending
the preceding type-based logic for updating the `free_symbols` set.