Fix: delete array moved to heap (backport to `v1/maintenance`) (#2135)
# Description

If arrays have storage type Register and are bigger than
`max_stack_array_size`, they are moved to the heap. In that case, since
allocation is now dynamic, the array also has to be de-allocated again.

This issue is also present in mainline DaCe. This PR is a backport of
[insert other PR here].

Co-authored-by: Roman Cattaneo <1116746+romanc@users.noreply.github.com>