Specialize scalar utility function (#2139)
ScalarToSymbol + SDFG.specialize breaks too often, and on f2dace, I
wrote this utility function to directly specialize a scalar to a
constant.

It does not fail when the scalar is used as a dynamic map connector and
used on interstate edges.

(I am not sure if I need to cover the case where the scalar name is the
subset of another array name)
(Btw. maybe we should add warnings, if a data descriptor name is subset
of another data descriptor because our string replacement functions can
create problems there)