Made `Config` Thread Safe (#2126)
This PR solves [issue#2125](https://github.com/spcl/dace/issues/2125),
which leads to problems with the configuration object in the presence of
multiple threads.
While DaCe (the Python project) was never designed to be used in multi
threaded context, it is actually used in such a way, for example inside
the tests.

The solution of the PR is to store the configuration inside a thread
local, aka. `threading.local`, variable.
This has the effect that every thread has its own set of configuration.
There is, however, a slight change in behaviour, as now the
configuration is initialized with the default setting when a new thread
is spawned.
Before, due to the sharing to the configuration, it looked like the new
thread "inherited" the configuration from its parent, but that was an
error anyway.