Safe call (#2163)
This PR adds a new `safe_call()` function to SDFGs, which runs the SDFG
in a seperate process in order to avoid crashing the main thread, when
the execution crashes (e.g. due to a segmentation fault). If the
execution crashes, `safe_call()` raises an exception instead, allowing
the user to handle it automatically.

---------

Co-authored-by: Philipp Schaad <schaad.phil@gmail.com>