Fixed `consolidate_edges_scope()` (#2184)
The function was assuming that
`inner_edge_starting_at_map_entry_node.data.data` does always refer to
the data that is outside the Map.
However, in case `inner_edge_starting_at_map_entry_node.dst` is an
AccessNode this is not necessarily the case.
This also leads to some issues with with `subset` vs. `other_subset`.

Note:
This PR also hit an issue in Memlet propagation that is not fixed, but
has the same root cause.