Removed that stupid sclar return value feature that CAN NOT WORK.
However, I saw that it also, under the hood sometimes tests if the argument is a pyobject.
Since that thing is a pointer it is possible and I implemented it for that.
But it was again not implemented properly, since for the case when the return value is passed as a regular argument, it was not checking that, only for managed return values.