MapFusion: Multiple Consumer of Second Map (#2157)
Before `MapFusionVertical` only allowed for one connection between an
intermediate (between the two Maps) and the second Map.
This PR changes the behaviour and now this case is allowed.
This also required an update of the function that checks if an
`AccessNode` was shared or not.
Which was refined to take this case into account.

It is important that the case where the first Map has multiple
connections to the intermediate is still not allowed.
However, in the longer run, this should be enabled.